Don't treat no changes as an error
When the change set contains no changes CFN returns an error. However, this is not an error state. When scripting with Cuffsert it complicates things a lot that no changes results in the tool exiting with a non-zero value. Either you have to ignore errors completely, or capture the output of the command and look for the "didn't contain changes" string – which is no fun at all in Bash.

I haven't been able to think of a situation where I want Cuffsert to exit with a non-zero value when there are no changes. If someone can come up with one it would probably be possible to add a flag that enabled a mode that treated no changes as an error state – but I would argue that the change I'm proposing here should be the default and that flag should be added when needed.