Add base64 and logger as a development dependencies
Since Ruby 3.4, the gem `base64` is a "bundled" gem, instead of a "default" gem, and needs to be added to the Gemfile.

It is used in the `aws-sdk-cloudformation` gem, but they have not added it as a dependency to not break running older versions of Ruby, therefore we don't add it as a runtime dependency here either.

Similar situation with `logger` which will be removed in Ruby 3.5.